Upgrade kotlin and compose related dependencies
This commit updates Kotlin to version 1.9.22 and compose related
dependencies.

It also updates the Kotlin compiler options in the build.gradle.kts
files and adds the compose-compiler plugin to the libs.versions.toml
file.

Additionally, two lint checks related to compose are disabled in the
app/build.gradle.kts file.